% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcqol_raw}
\alias{pcqol_raw}
\title{Example PC-QoL data}
\format{
## `pcqol_raw`
A data frame with	7 rows of 6 columns:
\describe{
  \item{upset}{You were upset because of your child's cough}
  \item{awake}{You were awakened during the night because of your child's cough}
  \item{leave}{You were worried/ concerned about leaving your child with other because of the cough}
  \item{leada}{You were worried/ concerned about your child being able to lead a normal life}
  \item{scare}{You were scared because of your child's cough}
  \item{overp}{You felt you were being overprotective because of your child's cough}
}
}
\source{
internal
}
\usage{
pcqol_raw
}
\description{
Made up data to see how the package works
}
\keyword{datasets}
