% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqol_raw2.R
\docType{data}
\name{pcqol_raw2}
\alias{pcqol_raw2}
\title{Example PC-QoL data - non-conforming variables names}
\format{
## `pcqol_raw2`
A data frame with	7 rows of 6 columns:
\describe{
  \item{ups}{You were upset because of your child's cough}
  \item{awa}{You were awakened during the night because of your child's cough}
  \item{lwo}{You were worried/ concerned about leaving your child with other because of the cough}
  \item{lnl}{You were worried/ concerned about your child being able to lead a normal life}
  \item{sca}{You were scared because of your child's cough}
  \item{ove}{You felt you were being overprotective because of your child's cough}
}
}
\source{
internal
}
\usage{
pcqol_raw2
}
\description{
Made up data to see how the package works - used to test the recode function
}
\keyword{datasets}
