% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqol_recode.R
\name{pcqol_recode}
\alias{pcqol_recode}
\title{Recode PC-QoL data}
\usage{
pcqol_recode(
  data,
  colnames = c(upset = "upset", scare = "scare", overp = "overp", leave = "leave", leada
    = "leada", awake = "awake"),
  replace_cols = FALSE
)
}
\arguments{
\item{data}{Input data. should be either a `data.frame` or tibble object.}

\item{colnames}{A list contain the name bindings for each PC-QoL item}

\item{replace_cols}{replace columns or create new columns?
* `TRUE`: columns will be replaced
* `FALSE`(default): new columns will be appended on to the `data.frame`}
}
\value{
A `data.frame` with each of the recoded fields, either where columns
  replaced by recoded, or in addition to the
}
\description{
`pcqol_recode()` recodes raw PC-QoL data in preparation for estimation of
health utilities. It is intended as a helper for `pcqol_utilities()`, but is
available for use on its' own if the user wishes.
}
\examples{
# replacing cols
pcqol_recode(pcqol_raw, replace_cols = TRUE)
# vs not replace cols
pcqol_recode(pcqol_raw, replace_cols = FALSE)
}
