% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqol_utilties.R
\name{pcqol_utilities}
\alias{pcqol_utilities}
\title{Estimate utilities}
\usage{
pcqol_utilities(
  data,
  recode = TRUE,
  colnames = c(upset = "upset", scare = "scare", overp = "overp", leave = "leave", leada
    = "leada", awake = "awake"),
  value_set = "roberts2024",
  return_df = TRUE
)
}
\arguments{
\item{data}{A data frame containing fields `upset`, `scare`, `overp`, `leave`, `leada`, `awake`, or fields specified in `colnames`.}

\item{recode}{Do the columns need to be recoded, from the 1-7 Likert scale, to the 1-4 Likert scale.}

\item{colnames}{A named character vector of fields of interest.}

\item{value_set}{A character from which value set should utilities be estimated.}

\item{return_df}{Logical. should the calculated utilities be appended to the originally input data frame.}
}
\value{
A tibble of utility values, with or without original data.
}
\description{
Estimate utilities
}
\examples{
# with a simple data frame
pcqol_utilities(pcqol_raw)
# a data frame with non-standard column names
pcqol_utilities(pcqol_raw2,
colnames = c(
  upset = "ups",
  scare = "sca",
  overp = "ove",
  leave = "lwo",
  leada = "lnl",
  awake = "awa"))
}
