% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqolutilities-package.R
\docType{package}
\name{pcqolutilities-package}
\alias{pcqolutilities}
\alias{pcqolutilities-package}
\title{pcqolutilities: PC-QoL Utilities}
\description{
Recode PC-QoL responses and estimate health utilities for responses.
}
\keyword{internal}
