% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqolutilities_shiny.R
\name{run_pcqol_utilities_app}
\alias{run_pcqol_utilities_app}
\title{Launch the PCQoL Utilities Shiny App}
\usage{
run_pcqol_utilities_app()
}
\value{
A shiny app for quick and easy estimation of PC-QoL utilities.
}
\description{
Launch the PCQoL Utilities Shiny App
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
run_pcqol_utilities_app()
\dontshow{\}) # examplesIf}
}
