% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqolutilities-package.R
\docType{package}
\name{pcqolutilities-package}
\alias{pcqolutilities}
\alias{pcqolutilities-package}
\title{pcqolutilities: Estimate Parent-Proxy Child Cough Specific Quality of Life Utilities}
\description{
Recode Parent-Proxy Child CoughSpecific Quality of Life questionnaire (PC-QoL) responses and estimate their respective health utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrob95/pcqolutilities}
  \item Report bugs at \url{https://github.com/jrob95/pcqolutilities/issues}
}

}
\author{
\strong{Maintainer}: Jack Roberts \email{jack.roberts@connect.qut.edu.au}

}
\keyword{internal}
