% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcqolutilities-package.R
\docType{package}
\name{pcqolutilities-package}
\alias{pcqolutilities}
\alias{pcqolutilities-package}
\title{pcqolutilities: Estimate Parent-Proxy Child Cough Specific Quality of Life Utilities}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Recode Parent-Proxy Child Cough-specific Quality of Life questionnaire (PC-QoL) responses and estimate their respective health utilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jrob95/pcqolutilities}
  \item Report bugs at \url{https://github.com/jrob95/pcqolutilities/issues}
}

}
\author{
\strong{Maintainer}: Jack Roberts \email{jack.roberts@connect.qut.edu.au} (\href{https://orcid.org/0000-0003-3519-4725}{ORCID})

Authors:
\itemize{
  \item Rex Parsons \email{rex.parsons94@gmail.com} (\href{https://orcid.org/0000-0002-6053-8174}{ORCID})
}

}
\keyword{internal}
